Make toolchain-desktop-macos depend on fetch-nss-artifact [ci full]
The `toolchain-desktop-macos` task implicitly depends on the
`fetch-nss-artifact` task, as it is using the index here:
https://github.com/mozilla/application-services/blob/41ecfc79dd1ea3c5962967707e11e0497a7e7703/libs/build-nss-desktop.sh#L61

However, it doesn't list the task as a dependency. This means that the
`fetch-nss-artifact` task never gets pulled into the graph and needs to be
run manually anytime it expires or is updated.

By adding this as a `fetch` to the `toolchain-desktop-macos`, it:

1. Sets up the dependency link in Taskcluster
2. Automatically downloads + extracts the artifact (so we can cleanup
   the build script a bunch)

+++ libs/build-nss-desktop.sh
@@ -58,11 +58,16 @@ fi
 # https://github.com/mozilla/application-services/issues/962
 if [[ "${CROSS_COMPILE_TARGET}" =~ "darwin" ]]; then
   #From https://firefox-ci-tc.services.mozilla.com/tasks/index/app-services.cache.level-3.content.v1.nss-artifact/latest
-  curl -sfSL --retry 5 --retry-delay 10 -O "https://firefox-ci-tc.services.mozilla.com/api/index/v1/task/app-services.cache.level-3.content.v1.nss-artifact.latest/artifacts/public/dist.tar.bz2"
-  SHA256="15e8cab63a4e0484564231a24d23d8be349217ba6add53b56609c339bb306829"
-  echo "${SHA256}  dist.tar.bz2" | shasum -a 256 -c - || exit 2
-  tar xvjf dist.tar.bz2 && rm -rf dist.tar.bz2
-  NSS_DIST_DIR=$(abspath "dist")
+  if [[ "${MOZ_AUTOMATION}" == "1" ]]; then
+    # run-task has already downloaded + extracted the dependency
+    NSS_DIST_DIR="${MOZ_FETCHES_DIR}/dist"
+  else
+    curl -sfSL --retry 5 --retry-delay 10 -O "https://firefox-ci-tc.services.mozilla.com/api/index/v1/task/app-services.cache.level-3.content.v1.nss-artifact.latest/artifacts/public/dist.tar.bz2"
+    SHA256="15e8cab63a4e0484564231a24d23d8be349217ba6add53b56609c339bb306829"
+    echo "${SHA256}  dist.tar.bz2" | shasum -a 256 -c - || exit 2
+    tar xvjf dist.tar.bz2 && rm -rf dist.tar.bz2
+    NSS_DIST_DIR=$(abspath "dist")
+  fi
 elif [[ "$(uname -s)" == "Darwin" ]] || [[ "$(uname -s)" == "Linux" ]]; then
   "${NSS_SRC_DIR}"/nss/build.sh \
     -v \
